Refs #238 Refactored logging initialization
Intended to clearly separate logging initialization into a
pluggable component.
