+++ conf/yawg-defaults.conf
@@ -47,4 +47,4 @@ fi
 # Informational.
 #
 YAWG_PRODUCT_NAME="Yawg"
-YAWG_COPYRIGHT_HEADER="Copyright (c) 2017 Yawg project contributos, All Rights Reserved"
+YAWG_COPYRIGHT_HEADER="Copyright (c) 2017 Yawg project contributors, All Rights Reserved"
