+++ devtools/bin/build-sync-version
@@ -1,7 +1,7 @@
 #!/bin/bash
 ###########################################################################
 #
-# Copyright (c) 2016 Yawg project contributors.
+# Copyright (c) 2016-2017 Yawg project contributors.
 #
 #
 # Updates the project version in all pom.xml files to match the
@@ -47,7 +47,7 @@ function main () {
 
     yawgCheckForTools \
         "${_xsltprocTool}"
-    
+
     updateVersion
 }
