--- doc/content/Developers/index.html
@@ -1,15 +1,0 @@
-<html>
-  <head>
-    <title>Developers</title>
-  </head>
-
-  <body>
-
-    <p>Here you will find information intended for the developers of
-    the Yawg project.</p>
-
-    <p><a href="CreatingRelease.html">How to create a release of the
-    Yawg software.</a></p>
-
-  </body>
-</html>
