+++ modules/cli/src/main/java/com/varmateo/yawg/cli/InfoPrinter.java
@@ -11,8 +11,8 @@ import java.text.MessageFormat;
 
 import com.varmateo.yawg.YawgInfo;
 import com.varmateo.yawg.cli.BakerCliOptions;
-import com.varmateo.yawg.cli.util.CliException;
-import com.varmateo.yawg.cli.util.CliInfoPrinter;
+import com.varmateo.yawg.cli.CliException;
+import com.varmateo.yawg.cli.CliInfoPrinter;
 
 
 /**
