--- modules/cli/src/main/java/com/varmateo/yawg/cli/util/package-info.java
@@ -1,10 +1,0 @@
-/**************************************************************************
- *
- * Copyright (c) 2016 Yawg project contributors.
- *
- **************************************************************************/
-
-/**
- * Utility classes for parsing command line arguments.
- */
-package com.varmateo.yawg.cli.util;
