+++ modules/pom.xml
@@ -47,9 +47,9 @@ Copyright (c) 2015-2017 Yawg project contributors.
     <maven-jxr-plugin.version>2.5</maven-jxr-plugin.version>
     <maven-pmd-plugin.version>3.7</maven-pmd-plugin.version>
     <maven-project-info-reports-plugin.version>2.8</maven-project-info-reports-plugin.version>
+    <maven-site-plugin.version>3.4</maven-site-plugin.version>
     <maven-surefire-plugin.version>2.18.1</maven-surefire-plugin.version>
     <maven-surefire-report-plugin.version>2.18.1</maven-surefire-report-plugin.version>
-    <maven-site-plugin.version>3.4</maven-site-plugin.version>
     <mockito-core.version>2.2.27</mockito-core.version>
     <yamlbeans.version>1.09</yamlbeans.version>
